editor: add `:edit` command (opens page with text box) [feat][wip]
Add a new `:edit` command that opens a local HTML page with a text
box.

The goal is to implement a simple in-browser text editor, for taking
notes while browsing the web.

This is a WIP. Right now it is just a text box with no CSS styling and
no ability to read/save files on the local filesystem.